#25: Quickly fix norm + Update coding style for much more clearance
