+++ source/common/microkernel_architecture/microkernel_modules/instructions/AInstruction.hpp
@@ -1,0 +1,26 @@
+#ifndef AINSTRUCTION_HPP_
+# define AINSTRUCTION_HPP_
+
+# include "AInstructionBusClient.hpp"
+
+class AInstruction
+{
+private:
+    AInstructionBusClient *_transmitter;
+    AInstructionBusClient::eClientId _receiver;
+    int _instructionId;
+
+public:
+    AInstruction(AInstructionBusClient *, AInstructionBusClient::eClientId, int);
+    virtual ~AInstruction();
+
+    void setTransmitter(AInstructionBusClient *);
+    void setReceiver(AInstructionBusClient::eClientId);
+    void setInstructionId(int);
+
+    AInstructionBusClient *getTransmitter() const;
+    AInstructionBusClient::eClientId getReceiver() const;
+    int getInstructionId() const;
+};
+
+#endif // !AINSTRUCTION_HPP_
